l1-gemini/pyproject.toml
[project]
name = "l1-gemini"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-genai>=1.59.0",
    "pillow>=12.1.0",
]
